Added a open file Window
Added an OpenFileDialog window for finding the GVS file
Changed the while(true) main loop to not use tick counts, but just sleep
for 200ms.
Added code so the program will now exit if the file path to the selected
file is not valid.

Added code so the program will print an error when the GVS is invalid.
The program will exit on next keypress.